Add files via upload
These files have the adjustPathForServerVsLocal function that, as the name suggests, adjusts the file paths for the EC2 server and compares them to the paths of the local machine. We were running into issues with the ./aws_config or similar files that were not being properly referenced on the EC2 server. So instead of hardcoding the paths locally then uploading to EC2, we have an automated version that can be used fairly easily.